Restrict S/MIME key validity to five years
This is to make somehow sure that the certificate is no valid beyond the
validity of the root certificate.  However, there is still no hard check in the
signing process and the fact whether the validity of the end user certificate is
not after the validity of the root certificate has to be checked by hand.